+++ src/Sentry/Extensibility/StringStackTraceFactory.cs
@@ -1,16 +1,25 @@
-using System;
-using System.Collections.Generic;
+using Sentry.Infrastructure;
 
 namespace Sentry.Extensibility;
+
+#if NET8_0_OR_GREATER
 
 /// <summary>
 /// A rudimentary implementation of <see cref="ISentryStackTraceFactory"/> that simply parses the
 /// string representation of the stack trace from an exception. This lacks many of the features
-/// off the full <see cref="SentryStackTraceFactory"/>, however it may be useful in AOT compiled
+/// off the full <see cref="SentryStackTraceFactory"/>. However, it may be useful in AOT compiled
 /// applications where the full factory is not returning a useful stack trace.
-/// <remarks>SimpleStackTraceFactory is currently EXPERIMENTAL.</remarks>
+/// <remarks>
+/// <para>
+/// This class is currently EXPERIMENTAL
+/// </para>
+/// <para>
+/// This factory is designed for AOT scenarios, so only available for net8.0+
+/// </para>
+/// </remarks>
 /// </summary>
-public partial class SimpleStackTraceFactory : ISentryStackTraceFactory
+[Experimental(DiagnosticId.ExperimentalFeature)]
+public partial class StringStackTraceFactory : ISentryStackTraceFactory
 {
     private readonly SentryOptions _options;
     private const string FullStackTraceLinePattern = @"at (?<Module>[^\.]+)\.(?<Function>.*?) in (?<FileName>.*?):line (?<LineNo>\d+)";
@@ -41,15 +50,15 @@ public partial class SimpleStackTraceFactory : ISentryStackTraceFactory
 #endif
 
     /// <summary>
-    /// Creates a new instance of <see cref="SimpleStackTraceFactory"/>.
+    /// Creates a new instance of <see cref="StringStackTraceFactory"/>.
     /// </summary>
     /// <param name="options">The sentry options</param>
-    public SimpleStackTraceFactory(SentryOptions options)
+    public StringStackTraceFactory(SentryOptions options)
     {
         _options = options;
     }
 
-    /// <inheritdoc cref="ISentryStackTraceFactory.Create"/>
+    /// <inheritdoc />
     public SentryStackTrace? Create(Exception? exception = null)
     {
         _options.LogDebug("Source Stack Trace: {0}", exception?.StackTrace);
@@ -57,8 +66,7 @@ public partial class SimpleStackTraceFactory : ISentryStackTraceFactory
         var trace = new SentryStackTrace();
         var frames = new List<SentryStackFrame>();
 
-        var newlines = new[] { Environment.NewLine };
-        var lines = exception?.StackTrace?.Split(newlines, StringSplitOptions.RemoveEmptyEntries) ?? [];
+        var lines = exception?.StackTrace?.Split(Environment.NewLine, StringSplitOptions.RemoveEmptyEntries) ?? [];
         foreach (var line in lines)
         {
             var fullMatch = FullStackTraceLine.Match(line);
@@ -94,6 +102,9 @@ public partial class SimpleStackTraceFactory : ISentryStackTraceFactory
         }
 
         trace.Frames = frames;
-        return trace;
+        _options.LogDebug("Created {0} with {1} frames.", "StringStackTrace", trace.Frames.Count);
+        return trace.Frames.Count != 0 ? trace : null;
     }
 }
+
+#endif